#42 refactor: MyPage 수정 - mypage.js 라우팅 수정 - 헤더 제거 - 미사용 InterestPage 제거
