--- frontend-set/src/pages/mypage/InterestPage.vue
@@ -1,7 +1,0 @@
-<template>
-  <h2>관심 페이지 (마이페이지 - 하트)</h2>
-</template>
-
-<script setup></script>
-
-<style scoped></style>
